+++ src/components/NewTable/index.ts
@@ -1,0 +1,1 @@
+export * from './NewTable';
