+++ src/components/Table/Table.spec.tsx
@@ -1,24 +1,27 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import { Table } from './Table';
-import { columns, data } from './data';
+import { rows, structure } from './data';
 
-describe('Table', () => {
+describe('Table Deprecated', () => {
   it('should render columns and rows properly', () => {
-    render(<Table columns={columns} data={data} />);
+    render(<Table rows={rows} structure={structure} />);
 
-    expect(screen.getAllByText('pending').length).toBe(1);
-    expect(screen.getAllByText('failed').length).toBe(5);
-    expect(screen.getAllByText('success').length).toBe(4);
+    // rows
+    expect(screen.getAllByText('data 1').length).toBe(5);
+    expect(screen.getAllByText('data 2').length).toBe(2);
+    expect(screen.getAllByText('data 3').length).toBe(3);
+    expect(screen.getAllByText('data 4').length).toBe(2);
 
-    expect(screen.getByText('user0@email0.com')).toBeVisible();
+    // renders child components
+    expect(screen.getByText('React node')).toBeVisible();
   });
   it('should call the onClick callback when a row is clicked', () => {
     const mockOnClick = jest.fn();
-    render(<Table columns={columns} data={data} onClick={mockOnClick} />);
+    render(<Table rows={rows} structure={structure} onClick={mockOnClick} />);
 
     fireEvent.click(screen.getByTestId('row-0'));
 
-    expect(mockOnClick).toHaveBeenCalled();
+    expect(mockOnClick).toHaveBeenCalledWith(rows[0]);
   });
 });
