+++ src/components/Table/data.tsx
@@ -1,57 +1,52 @@
-import { ColumnDef } from '@tanstack/react-table';
 import React from 'react';
-import { Button } from '../Button';
+import { TableStructure } from './types';
 
-type Payment = {
-  id: string;
-  amount: number;
-  status: 'pending' | 'processing' | 'success' | 'failed';
-  email: string;
+export const structure: TableStructure = {
+  header: [
+    {
+      title: 'Col 1',
+      name: 'col_1',
+      key: 'col_1',
+      width: 2,
+      mobile_width: 3,
+      show_title: true,
+      show_on_hover: false,
+    },
+    {
+      title: 'Col 2',
+      name: 'col_2',
+      key: 'col_2',
+      width: 1,
+      mobile_width: 1,
+      show_title: true,
+      show_on_hover: false,
+    },
+    {
+      title: 'Col 3',
+      name: 'col_3',
+      key: 'col_3',
+      width: 1,
+      mobile_width: 1,
+      show_title: true,
+      show_on_hover: false,
+    },
+    {
+      title: 'Col 4',
+      name: 'col_4',
+      key: 'col_4',
+      width: 1,
+      mobile_width: 3,
+      show_title: true,
+      show_on_hover: false,
+    },
+  ],
 };
 
-export const columns: ColumnDef<Payment>[] = [
-  {
-    accessorKey: 'email',
-    header: 'email',
-    minSize: 200,
-    maxSize: 300,
-  },
-  {
-    accessorKey: 'status',
-    header: 'Status',
-  },
-  {
-    accessorKey: 'random',
-    header: 'Random',
-  },
-  {
-    accessorKey: 'amount',
-    header: () => <div>Amount</div>,
-    cell: ({ row }) => {
-      const amount = parseFloat(row.getValue('amount'));
-      const formatted = new Intl.NumberFormat('en-US', {
-        style: 'currency',
-        currency: 'USD',
-      }).format(amount);
-
-      return <div>{formatted}</div>;
-    },
-  },
-  { id: 'action', enablePinning: true, cell: () => <Button>View</Button> },
+export const rows = [
+  { key: '1', col_1: <div>React node</div>, col_3: 'data 3' },
+  { key: '2', col_1: 'data 1', col_2: 'data 2', col_4: <div>data 4</div> },
+  { key: '3', col_1: 'data 1', col_3: 'data 3' },
+  { key: '4', col_1: 'data 1', col_3: 'data 3' },
+  { key: '5', col_1: 'data 1', col_4: 'data 4' },
+  { key: '6', col_1: 'data 1', col_2: 'data 2' },
 ];
-
-export const data: Payment[] = [...new Array(10)].map((_, index) => ({
-  id: index.toString(),
-  amount: index + 100,
-  email: `user${index}@email${index}.com`,
-  status: index === 0 ? 'pending' : index % 2 === 0 ? 'success' : 'failed',
-  random: 'hola',
-}));
-
-export const dataPagination: Payment[] = [...new Array(1005)].map((_, index) => ({
-  id: index.toString(),
-  amount: index + 100,
-  email: `user${index}@email${index}.com`,
-  status: index === 0 ? 'pending' : index % 2 === 0 ? 'success' : 'failed',
-  random: 'hola',
-}));
