+++ src/components/index.ts
@@ -15,8 +15,8 @@ export { default as RadioGroup } from './RadioGroup';
 export * from './Select';
 export * from './Stepper';
 export { default as Switch } from './Switch';
-export * from './TableDeprecated/Table';
-export * from './Table';
+export * from './Table/Table';
+export * from './NewTable';
 export { default as Tabs } from './Tabs';
 export { default as TextArea } from './TextArea';
 export { TextField, MuiTextField } from './TextField';
